frontend/requirements.txt
streamlit
supabase
stripe
python-dotenv
requests
pandas